Learning about Interpreters and Compilers.
In this project, I plan to create a working toy programming language
such that it can lex, parse and execute its supposed source files in the
/src directory. My guide in this interpreter/compiler design rabbithole
is the book `Crafting Interpreters` by `Robert Nystrom`.

The /src directory files contain a miniature version of the syntax of
the language, named Lox. More sysntax definitions will be added from
operators to other sections of the language.