Added functions and closures to Lox
It keeps getting better and better, the function definitions for Lox are
just a block with some injected variables which are the parameters with
the initial values being the arguments. Boooyaaaaa!

There was a massive error in env, you know how python treats empty
containers as false and true otherwise, yep, to check if an environment
has a parent, i used a bool check on it which sometimes returns false if
the parent env is empty therefore, I create a new env, skipping the
empty parent therefore, other code that uses the originally empty env
would fill it up and i wouldn't be able to use that env's resources,
nasty. Hard to track down, but yah buoi over here found it. Booooyaaaa!