Added break keyword and fixed some bugs.
The break keyword was easy to add, same implementation as return
keyword. For the continue, well, not so easy, for a while loop it works
like a charn, but a desugared for-while-loop was hell due to the last
statement that had to run to increment or do some stuff in which the
continue has to update before the next iteration, if skipped, may cause
an infinite loop. With this, I tried changing a few things here and
there but that quickly backfired as it needed me to change a lot and at
the end of all that, a simple change that was suppossed to work, some
how lost the resolution for the variable to be incremented, HELLL!!
I blame this to poor design, resolution and variable lookups were poorly
implemented, but, this is just a simple language to learn about
compilers. Oi.