Added classes, instances, fields and methods.
Best chapter ever, I've finally understood a little of how object
oriented programming languages work. Not thoroughly but it WORKS!!!
Classes and instances are simply data pools, classes store methods while
instances store fields, accessing a method via an instance will look for
the method from the class in which it will be bound to the method and
return a simple function in which the 'this' within resolves to the
bound instance, accessing them directly from the class return unbound
methods that expect one argument, the instance, which will return a
bound function to the passed instance which can then be used as if it
was accessed from the instance directly.
We also have a constructor, init. If it is a method, it can be used to
construct new Instances, if not, it is then a class variable.
Classes also support class variables, later I will add class methods and
static methods. Simply, a static method will be a method bound to None
and class methods, you guessed it, the class itself. Amazing.
For now, just access the method and pass the class itself as an instance
therefore class methods, and attach normal functions to the class and
you will have static methods since they are not allowed to have this in
them and a missing field in the instance is searched in the class hence,
instances can call them as they also resolve to functions, normal ass
functions. (^_^)