Added a scanner/lexer.
I never thought generators were possible in c. I mean we did create
somesort of generators, returns the next Token from the source code when
asked to, it doesn't allocate memory for the tokens, they simply point
to the address of the source code where to start and the length of the
token. This is amazing stuff. And to think we cannot print a substring
of a string in c, well, I introduce to you `printf("%.*s", len, start)`
where length is the length of the substring and start is a `char *` to
the start of the substring in the original string. Truly amazing. I
think am falling in love with C. No, I'm definately loving C.