Implemented the compiler.
I'm gonna be honest, I didn't understand a thing in this chapter,
especially the Pratt's precedence algorithm. Maybe because I just
implemented it in C which is very verbose. I will reimplement it in
python in the project YAP but for now, I will simply read on it. I will
stick to this chapter till I feel comfortable with its contents. Though
it is amazing how the compilation/execution pipeline is taking form.