Read on pratt's algorithm and fixed some bugs
Last time I got stack on Pratt's algorithm, turns out it is quite simple
and elegant and can truly withstand any-fix operatos, be it pre-, post-,
in-, mix- and any other fix you have, bring it on. This is ofcourse
theoretical and I'm yet to test it out, I will try to incoporate this
experiment into yap. The bugs part is mainly on comments, for some weird
reason, comments were not parsed correctly, fixed it and also added
multiline comments as in jlox's.