Added some operators and value types
This is not exactly a type system but the implementation of value types
is really amazing, the values being unions, something I have never used
in all my time using C, the tagging part, genius idea, I saw something
like this in c++ metaprogramming, called tag dispatch, where there were
functions of similar signatures but one of its parameters was of a
specific type, either value parameter, template parameter which was used
to differentiate all of the overloads, amazing. I have to look at type
systems, later though. For now we have implemented number, boolean and
nil types, strings are a little more delicate and it deserves its own
commit. For the operators, we have implemented comparrison, equality and
not operators, they work amazingly and on top of that, I got to see C's
variable args in action from a format string similar to printf's calling
signature.