Added strings to lox and a basic garbage collector
Struct inheritance, have you ever haerd of it, other name, type punning,
It is quite powerful. With this feature, I see how people program
different applications using this simple barebone language C. I will
document about it in the code repo under C source files, get it there.
With this neat feature you can have an Value that can be anything where
the current type is tagged onto the value and casting it to the type
represented by the tag spans it to the required type, whoa. As for the
garbage collector, it is simply deleting allocated memory on vm_delete
where all values are being stored in a singly linked list.