Fixed one hell of a bug
I know C is known for the tricky memory handling canandrome but I never
new how serious it is. Like, just allocate memory and free it when
you're done with it, dead simple. Unfortunately that is not all, you
have to know where to store it, free it when you're done with it, and
rememver which ones were heap allocated and which ones are not. My last
commit about yesterday on this repo involved interning strings if I
remember correctly where I used the Table object as a hash set, the
problem occurred when I was growig the table which involved a few steps,
unfortunately I tripped along the was and ended up losing the original
entries memory address that I was to free before returning from the
function, this is where the mess is, I ended up freeing memory I did not
allocate hence an invalid pointer exit was raised my the OS hence
killing my program. It was hard to find this in this uncoordinated
jumble of a project. But a I did cook up a simple way to find the
invalid free call, just add a simple whence before the reallocate call
for free, I mean, use `__LINE__` macro in the FREE_ARRAY macro which
printed where the free was called from, sounds easy but I had to think
hard to get to this solution. All in all, bug fixed, clox back to life
and I can sleep peacefully tonight. NIGHT.