Added Local Variables elegantly.
I thought local variables would require the HashTable and some kind of
chaining but no, not at all. By using the stack, the vm one, we can use
it to store the local variables. We know that the overall effect of
statements on the stack is zero, but by letting var in scoped context
leave a value on the stack and use its index as the location marker then
anywhere that declared variable is used, we can use the OP_[SG]ET_LOCAL
<location> to instantly load a local variable, cleaning up is the
easiest bit, just pop from the stack as many variables as there were
defined in the exiting scope, neat. The variables stay there like
ghosts, since our parser produces really clean instructions, I mean, If
if wants a value then it executes a value emitting instruction then pop
it from the stack, cleanly, till now, all statements act this way except
for this scoped var statement but it also cleans up at the end of the
scope by popping all the declared vars, nice. I'm curious how we will
implement the closures, not sure but its gonna be lovely doing it.