Added Control Flow and Loops to Lox.
In this commit, lox has if-else, while and for statements implemented
using a few new instructions like JUMP_IF_FALSE, JUMP, and LOOP. The
difference being that JUMP_IF_FALSE jumps if the top value in the stack
is falsey, JUMP and JUMP_IF_FALSE jump forward by the provided offset
while LOOP jumps backwards. Using these four instructions, we were able
to implement the forementioned statements, it was exciting. The for loop
at fast was really intimidating, I mean, we parse the initializer,
condition, increment and body statements respectively but the insrement
has to run after each iteration (body) and the condition after the
initializer then the body again, convolution. Drawing a simple finite
state machine clarified everything, if every arrow is a jump instruction
to the instruction to run next then we can easily code it and wala, for
loop.

Drawing:

                        +------------------> stop
                        | False
	start -------> cond <------- incr
	                |              |
			| True         | ^
			|              |
		      body ----------->+

	<init>
	<cond>
	JUMP_IF_FALSE <end_of_loop>
	POP
	JUMP <body>
	<incr>
	LOOP <cond>
	<body>
	JUMP <incr>
	<end_of_loop>
	POP

Example:
	for (var count = 100; count > 0; count = count - 10)
		print count * count;

	1  CONSTANT      100
	2  SET_LOCAL     count
	3  GET_LOCAL     count
	4  CONSTANT      0
	5  GREATER
	6  JUMP_IF_FALSE 18      -----+<---+
	7  POP                        |    |
	8  JUMP          13      ---------------------+
	9  GET_LOCAL     count   <--------------+     |
	10 CONSTANT      10           |    |    |     |
	11 SET_LOCAL     count        |    |    |     |
	12 LOOP          6       ----------+<---------+
	13 GET_LOCAL     count        |         |
	14 GET_LOCAL     count        |         |
	15 MULTIPLY                   |         |
	16 PRINT                      |         |
	17 LOOP          9      ----------------+
	18 POP                  <-----+

For the diagram, I've used the actual instruction number instead of the
offset to make things simpler, remember, LOOP jumps backwards while JUMP
forward, weird sentence.
This is just unbelievable.