Added partial support for closures
The upvalues trick used to solve closures as used in Lua is quite
tricky, like the pratt's expression parser algorithm, I needed some time
to process this chapter, I've not yet understood it but so far so good.
If I've gotten it right then what we have done so far is retrieving
variables from the stack for a non-returned function, like OP_LOOP,
OP_SET_UPVALUE/OP_GET_UPVALUE are referencing the stack backwards from
the closures local window but instead of using n index, the upvalues
have the address of the slot where the captured variable lives. In code,
this is messy and hard to follow through but in theory it is as easy as
I have just made it. So far, I've implemented closures if the parent(s)
functions scopes have not returned, ie, a closure only works if it is
called where it was defined otherwise it might misbehave due to the
stack being overwritten as the program goes on. The next section of this
closure endeavour is about persisting that slot, using the HEAP. This
way, even if the parent returns, the captured locals of it will still
liveon for the closure to behave expectedly, Bonus, if more than
closures leave the parent while capuring the ame variable, then they
will continue catching the same variable in the heap. I do not know how
this will be implemented but I'm eager to see the magic. NIGHT.