Completed the implementation of closures and upvalues
Upvalues, isn't there a much better, easier solution to this. Like,
analyzing the code then determining which locals have been captured and
storing all this info into some symbol table of infomation database then
producing instructions that fits the infomation using simpler mechanics
would be better. My conclusion is that this book posses simple ideas
which I love about the book but the simpleness is lost in the code
implementation. The code just obscures the elegance of the algorithms.
Though this codebase is heavily mismanaged and right now some files have
over 600 lines and that is really confusing to read and track. My idea
is that meaning should be part of the syntax, the code should hold its
own meaning. Like python generators, they have a yield keyword which
when used inside a function marks that function as a generator instead
of having a weird syntax for generator functions like in javascript
where a generator function is a normal function with an asterisk between
the function keyword and its name. Its good and stuff but then they have
the yield keyword. I think this comes down to greedy vs dynamic. Greedy
is stupid and it does get the job done somtimes its just that its not
always the best solution. For example, with the asterisk thing in
javascript, it does signify a generator function before we even start
parsing the functions body but now we end up with two ways of
identifying generator functions, via the presence of yield and the
asterisk, what a waste. For the dynamic solution, if we analyze the
functions body and collect all infomation about it like the presence of
the yield keyword then we would know that this is not a normal function
and hence produce the right kind of object, amazing. I read this greedy
vs dynamic thing and to say the truth, I thought it was an algorithm
thing but it is also how you process and synthesize a problem. Analysis
of a problem to get more info about it is what I would say about it.
Dynamic and all, NIGHT.