Fixed the Garbage Collector Bugs.
Inspite of the abortion of the GC, I sat down and run through the messy
code and patched a whole bunch of errors. Along the way I met segfault,
realloc invalid_pointer, hash function and intern string errors. The
most disturbing and hard t fix was the segfault, but I managed. The hash
function errors mostly are from the fact that my hash function and the
one presented by the author are pretty basic, as a fix, I made my hash
function produce a little bit bigger integers and also made the hash
value type bigger from 32-bit integer to 64, but to say the truth, most
errors were around the hash table since no resolution of global
variables led to a runtime error and I messed up with pointers to Value
and Object which led to SegFaults and Realloc errors. For the intern
strings, as I said with the failure to resolve keys in the table led to
creation of new strings which is a problem since, as an optimization, we
were using a strings address to know if it was equal to another and as
you might have guessed, this led to missing a value from the global
resolution which again led to the runtime error.
I have a problem ith the hash function since fixing it for some strings
leads to others not working but for now I've balanced it and it works
for all I know but as a fall back, I have commented a line that checks
two strings character by character to see if they are equal hence global
resolution works, but I hope to implement stronger hash functions later.
Today I've learned about debugging and I almost smashed my head against
the wall but I did not, I can now finish the GC Chapter and continue
with the walk of life, HappyDay.