Added methods and initilizers
I'm kinda getting used to this, the programming part of it, what
software is and how to think about problems and their solutions and I
have to say, I'm glad I stumbled upon this amazing book.

One: Programmming is an art, the language is the pen and the computer is
the canvas: ide. Drawing on it is writing down a solution to your
problem. The only difference with the other literal drawing art is that
your programs are alive.

Two: Programming is easy, it doesn't matter what language you're using,
its about using the given tools/features to describe your solution to some
problem, yes, some languages are easier to do that in while others are
just plain simple.

Three: Systems are engines, the gears, bolts, nuts and oil are akin to
functions, mixins, bound/partial objects and features. When you create a
system think of the operations abstractly and use interfaces; this seems
to produce beautiful and easy to maintain systems.

Otherwise, I've added methods declaration, call and resolution to c-lox
classes where the initializers are special methods; though I'm
definately not impressed with the overall design of c-lox c code but my
implementation of a toy language will be better.

All in all, a clean design makes a simple system: its not that python is
easy, rather that it is well designed.