Finally, Garbage Collector works gracefully.
The garbage collector has made me wanna kill my self. It is made of a
simple algorithm, mark-sweep, but damn, that is all lost how objects
move around in C-Lox. But hey, I've concurred this beast, at first I
gave up and thought of aborting it, then I took sometime, 6hrs, to track
down how objects live and move in C-Lox then patched a buttload of
errors which did let the GC run but still was buggy, then I gave up on
it in favor of finishing other topics in the book, and now that I'm done
with the book (except the last chapter that is about optimization for
later) I've revisited the old GC chapter and fixed a whole bunch of
error, added the dynamic GC invocation and watched it churn away a
simple GC oriented program in which it perfomed amazingly.

The main error, the hardest to catch was that on invoking the GC, it did
a sequence of operations: marking, blackening then deleting. But, the
deleting part lead to calling reallocate which than called the GC again
and this second call deleted a bunch of objects that were still
referenced by the first call which led to a seg fault when control
returned to the first call. BigBrainsSimon in the room.

Inspite of this amazing success, the GC still has some bugs mucking
around in there, I can feel it, but for now, I'm satisfied by the result
and shallow grace runs it has performed. More debugging later. Night.