Fixed the hash table and GC's main bug
I was wondering how a hash value can affect the Table and I concluded
that there must e a bug in the hash table somewhere. I was right, I have
a problem with making code look concise and fast; use less lines and
variables, this is where the bugs hid. In my brief `table_adjust_cap`
function that was in charge of adjusting the table's capacity, I lost
something as I was heavily reusing variables and let me tell you I could
not figure out what it was. I rewrote the function in a very readable
manner and even added comments then out of nowhere, the hash table works
perfectly and better yet, the GC also works even better and faster.

This but was tracked by my ealier technique of attaching WHENCE printer
to the memory macros which showed be that `table_adjust_cap` allocated
memory and before returning it raised a segfault sometimes and this was
the giveaway. With this, I could finally lay another SegFault beast.

Now, there is just one chapter that awaits thee; Optimization. I had
skimmed the chapter and I think more bugs will be caught as I finish it
of. AfterNoon.