Updated QNAN bits from 13 to 11
Finally, we're done with C version of lox but to be fair pylox and clox
are nothing like each other; they have very different structures and
models and I look forward to fixing that by recreating clox and making
the structure be very much like pylox. This will act as an exercise of
my understanding of the project this far and also hope to make it more
efficient and well designed compared to the current version.

To the new and beyond, gracious.